# [WIP] Enhance `pick_channels` and optimize channel selection code