… mne/io/tests/test_pick.py
… mne/viz/evoked.py
… mne/viz/misc.py
… mne/viz/topo.py
… mne/viz/utils.py