This PR gives more power to the `pick_channels` function, so it can be more widely used throughout the codebase. The resulting code is more DRY and performs better. The performance improvements will be more relevant if and when #2765 gets merged.

Take the following scenario:

    ch_names = ['CH%03d' % x for x in range(400)]
    info = mne.create_info(ch_names, 1000.)
    ch_sel = list(ch_names)  # select all channels

Common channel selection idiom 1:

    [info['ch_names'].index(ch) for ch in ch_sel]
    >>> 100 loops, best of 3: 2.87 ms per loop

Propose to change to:

    mne.pick_channels(info['ch_names'], include=ch_sel, order='include', strict=True)
    >>> 1000 loops, best of 3: 235 µs per loop

Common channel selection idiom 2:

    [info['ch_names'].index(ch) for ch in ch_sel if ch in info['ch_names']]
    >>> 100 loops, best of 3: 5.59 ms per loop

Propose to change to:

    mne.pick_channels(info['ch_names'], include=ch_sel, order='include')
    >>> 1000 loops, best of 3: 236 µs per loop

Common channel selection idiom 3:

    [ch for ch in ch_sel if ch in info['ch_names']]
    >>> 100 loops, best of 3: 2.75 ms per loop

Propose to change to:

    ch_names = set(info['ch_names'])
    [ch for ch in ch_sel if ch in ch_names]
    >> 10000 loops, best of 3: 49.9 µs per loop