@wmvanvliet 2016-02-02
yeah, this is a bad idea for various reasons.